add new evaluation report component on different route
