+++ src/components/manageFiles/ProcessDetail/ProcessDetail.js
@@ -4,7 +4,10 @@ import styles from './ProcessDetail.module.css'
 import { toast } from 'react-toastify';
 import { processReport } from '../../../httpCommon/httpCommon';
 import axios from 'axios';
+import { Link } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { processDetailReport } from '../../../utils/helpers';
+import { Button } from 'react-bootstrap';
 import {
   Chart as ChartJS, ArcElement, Tooltip, Legend, CategoryScale,
   LinearScale,
@@ -31,6 +34,11 @@ ChartJS.register(ArcElement, Tooltip, Legend, CategoryScale,
 
 const ProcessDetail = () => {
   const { ProcessDetail } = useSelector((state) => state.app);
+  const navigate = useNavigate();
+
+  const handleNavigate = () => {
+    navigate('/processes/evaluation-report');
+  };
   // const dispatch = useDispatch();
   // const { ProcessDetail } = useSelector((state) => state.app);
 
@@ -150,19 +158,12 @@ const ProcessDetail = () => {
             </div>
           </div>
 
-          <div className={styles.processName}>
-            <img src={greenImage} alt={'Green Image'} />
-            <div className={styles.processHeading}>
-              <h3>Evaluation Report:</h3>
-            </div>
+
+          <div className={`d-flex flex-column align-items-center justify-content-center`}>
+              <Button onClick={handleNavigate} variant="success" className="mb-4">
+                Click here to see Evaluation Report
+              </Button>
           </div>
-          <div className={styles.info_container}>
-            <EvaluationReportComponent />
-            {/* <TestChart /> */}
-          </div>
-
-
-
 
           {/* <h3 className={styles.Process_Title}>Steps :</h3> */}
 
@@ -340,208 +341,208 @@ export const pieChartData = {
   ]
 };
 
-const EvaluationReportComponent = () => {
-  const [reportData, setReportData] = useState(processDetailReport);
-  const { ProcessDetail } = useSelector((state) => state.app);
-  console.log("ProcessDetail", ProcessDetail._id)
+// const EvaluationReportComponent = () => {
+//   const [reportData, setReportData] = useState(processDetailReport);
+//   const { ProcessDetail } = useSelector((state) => state.app);
+//   console.log("ProcessDetail", ProcessDetail._id)
 
-  const normalityAnalysisData = {
-    labels: ['Actual Areas', 'Rectangle Area', 'Slope', 'Slope Percentage Deviation', 'Calculated Slope'],
-    datasets: [
-      {
-        label: 'Analysis Value',
-        data: 
-        [
-          reportData?.normality_analysis?.list1?.actual_areas || 3.42,
-          reportData?.normality_analysis?.list1?.rectangle_area || 3.73,
-          reportData?.normality_analysis?.list1?.slope[0] || 1.84, // assuming only one value in slope array
-          reportData?.normality_analysis?.list1?.slope_percentage_deviation || 0.25,
-          reportData?.normality_analysis?.list1?.calculated_slope || 1.84
-        ],
-        backgroundColor: '#FFCE56'
-      }
-    ]
-  };
+//   const normalityAnalysisData = {
+//     labels: ['Actual Areas', 'Rectangle Area', 'Slope', 'Slope Percentage Deviation', 'Calculated Slope'],
+//     datasets: [
+//       {
+//         label: 'Analysis Value',
+//         data: 
+//         [
+//           reportData?.normality_analysis?.list1?.actual_areas || 3.42,
+//           reportData?.normality_analysis?.list1?.rectangle_area || 3.73,
+//           reportData?.normality_analysis?.list1?.slope[0] || 1.84, // assuming only one value in slope array
+//           reportData?.normality_analysis?.list1?.slope_percentage_deviation || 0.25,
+//           reportData?.normality_analysis?.list1?.calculated_slope || 1.84
+//         ],
+//         backgroundColor: '#FFCE56'
+//       }
+//     ]
+//   };
 
-  const barChartData = {
-    labels: ['Mean', 'Median', 'Mode'],
-    datasets: [
-      {
-        label: 'Scores Statistics',
-        data: [
-          reportData?.central_tendencies?.normal_dist?.mergedMean || 5.0,
-           reportData?.central_tendencies?.normal_dist?.mergedMedian || 4.0,
-            reportData?.central_tendencies?.normal_dist?.mergedMode?.length || 4.0
-        ],
-        backgroundColor: [
-          'rgba(255, 99, 132, 0.6)',
-          'rgba(54, 162, 235, 0.6)',
-          'rgba(255, 206, 86, 0.6)'
-        ],
-        borderColor: [
-          'rgba(255, 99, 132, 1)',
-          'rgba(54, 162, 235, 1)',
-          'rgba(255, 206, 86, 1)'
-        ],
-        borderWidth: 1,
-      }
-    ]
-  };
+//   const barChartData = {
+//     labels: ['Mean', 'Median', 'Mode'],
+//     datasets: [
+//       {
+//         label: 'Scores Statistics',
+//         data: [
+//           reportData?.central_tendencies?.normal_dist?.mergedMean || 5.0,
+//            reportData?.central_tendencies?.normal_dist?.mergedMedian || 4.0,
+//             reportData?.central_tendencies?.normal_dist?.mergedMode?.length || 4.0
+//         ],
+//         backgroundColor: [
+//           'rgba(255, 99, 132, 0.6)',
+//           'rgba(54, 162, 235, 0.6)',
+//           'rgba(255, 206, 86, 0.6)'
+//         ],
+//         borderColor: [
+//           'rgba(255, 99, 132, 1)',
+//           'rgba(54, 162, 235, 1)',
+//           'rgba(255, 206, 86, 1)'
+//         ],
+//         borderWidth: 1,
+//       }
+//     ]
+//   };
 
-  const pieChartData = {
-    labels: ['Score 1', 'Score 2', 'Score 3', 'Score 4', 'Score 5', 'Score 6', 'Score 7'],
-    datasets: [
-      {
-        label: 'score list',
-        data: reportData?.score_list || [1, 3, 2, 10, 7, 4, 8], // Values from the 'score_list'
-        backgroundColor: [
-          '#FF6384',
-          '#36A2EB',
-          '#FFCE56',
-          '#8B4513',
-          '#98FB98',
-          '#20B2AA',
-          '#FF4500'
-        ],
-        hoverBackgroundColor: [
-          '#FF6384',
-          '#36A2EB',
-          '#FFCE56',
-          '#8B4513',
-          '#98FB98',
-          '#20B2AA',
-          '#FF4500'
-        ],
-        borderWidth: 1,
-      }
-    ]
-  };
+//   const pieChartData = {
+//     labels: ['Score 1', 'Score 2', 'Score 3', 'Score 4', 'Score 5', 'Score 6', 'Score 7'],
+//     datasets: [
+//       {
+//         label: 'score list',
+//         data: reportData?.score_list || [1, 3, 2, 10, 7, 4, 8], // Values from the 'score_list'
+//         backgroundColor: [
+//           '#FF6384',
+//           '#36A2EB',
+//           '#FFCE56',
+//           '#8B4513',
+//           '#98FB98',
+//           '#20B2AA',
+//           '#FF4500'
+//         ],
+//         hoverBackgroundColor: [
+//           '#FF6384',
+//           '#36A2EB',
+//           '#FFCE56',
+//           '#8B4513',
+//           '#98FB98',
+//           '#20B2AA',
+//           '#FF4500'
+//         ],
+//         borderWidth: 1,
+//       }
+//     ]
+//   };
 
-  let promotersScores = 0;
-  let passiveScores = 0;
-  let DetractorsScores = 0;
-  if (reportData && reportData.score_list) {
-    const scoreListLength = reportData.score_list.length;
-  
-    promotersScores = ((reportData.score_list.filter(score => score === 9 || score === 10).length) / scoreListLength) * 100;
-    passiveScores = ((reportData.score_list.filter(score => score === 7 || score === 8).length) / scoreListLength) * 100;
-    DetractorsScores = ((reportData.score_list.filter(score => score > 0 && score <= 6).length) / scoreListLength) * 100;
-  }
+//   let promotersScores = 0;
+//   let passiveScores = 0;
+//   let DetractorsScores = 0;
+//   if (reportData && reportData.score_list) {
+//     const scoreListLength = reportData.score_list.length;
 
-  const npsScoreDistributionData = {
-    labels: ['Detractors', 'Passives', 'Promoters'],
-    datasets: [
-      {
-        data: [DetractorsScores || 20, passiveScores || 10, promotersScores || 65], // Assuming percentages for each category
-        backgroundColor: ['#FF6384', '#36A2EB', '#FFCE56'],
-        hoverBackgroundColor: ['#FF6384', '#36A2EB', '#FFCE56']
-      }
-    ]
-  };
+//     promotersScores = ((reportData.score_list.filter(score => score === 9 || score === 10).length) / scoreListLength) * 100;
+//     passiveScores = ((reportData.score_list.filter(score => score === 7 || score === 8).length) / scoreListLength) * 100;
+//     DetractorsScores = ((reportData.score_list.filter(score => score > 0 && score <= 6).length) / scoreListLength) * 100;
+//   }
 
-  useEffect(() => {
-    console.log(ProcessDetail);
-    const fetchData = async () => {
-      try {
-        const requestBody = { process_id: 'abc0099986567abcd' };
-        const response = await axios.post('https://100035.pythonanywhere.com/evaluation/evaluation-api/?report_type=process', requestBody);
-        console.log("response", response.data.score_list)
-        setReportData(response.data);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      }
-    };
+//   const npsScoreDistributionData = {
+//     labels: ['Detractors', 'Passives', 'Promoters'],
+//     datasets: [
+//       {
+//         data: [DetractorsScores || 20, passiveScores || 10, promotersScores || 65], // Assuming percentages for each category
+//         backgroundColor: ['#FF6384', '#36A2EB', '#FFCE56'],
+//         hoverBackgroundColor: ['#FF6384', '#36A2EB', '#FFCE56']
+//       }
+//     ]
+//   };
 
-    fetchData();
-  }, []);
+//   useEffect(() => {
+//     console.log(ProcessDetail);
+//     const fetchData = async () => {
+//       try {
+//         const requestBody = { process_id: 'abc0099986567abcd' };
+//         const response = await axios.post('https://100035.pythonanywhere.com/evaluation/evaluation-api/?report_type=process', requestBody);
+//         console.log("response", response.data.score_list)
+//         setReportData(response.data);
+//       } catch (error) {
+//         console.error('Error fetching data:', error);
+//       }
+//     };
 
-  if (!reportData) {
-    return <Spinner animation="grow" variant="success" />;
-  }
+//     fetchData();
+//   }, []);
 
-  console.log("reportData", reportData)
-  // Process reportData and structure it for chart display
-  const barChartOptions = {
-    maintainAspectRatio: false,
-    responsive: false,  // Set to false to prevent resizing
-    // scales: {
-    //   y: {
-    //     beginAtZero: true
-    //   }
-    // },
-    height: "222px",    // Set your desired height
-    width: "450px"      // Set your desired width
-  };
-  return (
-    <div>
-      <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between' }}>
-        <div className={styles.processHeadingChart2}>
-          <h5>Normality Analysis Data:</h5>
-        </div>
+//   if (!reportData) {
+//     return <Spinner animation="grow" variant="success" />;
+//   }
 
-        <div className={styles.processHeadingChart2}>
-          <h5>Bar Chart for Central Tendencies:</h5>
-        </div>
-      </div>
+//   console.log("reportData", reportData)
+//   // Process reportData and structure it for chart display
+//   const barChartOptions = {
+//     maintainAspectRatio: false,
+//     responsive: false,  // Set to false to prevent resizing
+//     // scales: {
+//     //   y: {
+//     //     beginAtZero: true
+//     //   }
+//     // },
+//     height: "222px",    // Set your desired height
+//     width: "450px"      // Set your desired width
+//   };
+//   return (
+//     <div>
+//       <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between' }}>
+//         <div className={styles.processHeadingChart2}>
+//           <h5>Normality Analysis Data:</h5>
+//         </div>
 
-      <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between' }}>
+//         <div className={styles.processHeadingChart2}>
+//           <h5>Bar Chart for Central Tendencies:</h5>
+//         </div>
+//       </div>
 
-        <Bar data={normalityAnalysisData} options={barChartOptions} />
-        <Bar data={barChartData} options={barChartOptions} />
-      </div>
-      <br/>
+//       <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between' }}>
 
-      <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between' }}>
-        <div className={styles.processHeadingChart2}>
-          <h5>Pie Chart for Score List:</h5>
-        </div>
+//         <Bar data={normalityAnalysisData} options={barChartOptions} />
+//         <Bar data={barChartData} options={barChartOptions} />
+//       </div>
+//       <br/>
 
-        <div className={styles.processHeadingChart2}>
-          <h5>Pie Chart for NPS Score Distribution:</h5>
-        </div>
-      </div>
+//       <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between' }}>
+//         <div className={styles.processHeadingChart2}>
+//           <h5>Pie Chart for Score List:</h5>
+//         </div>
 
-      <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between' }}>
+//         <div className={styles.processHeadingChart2}>
+//           <h5>Pie Chart for NPS Score Distribution:</h5>
+//         </div>
+//       </div>
 
-        <Pie
-          data={pieChartData}
-          options={barChartOptions}
-        />
-        <Pie
-          data={npsScoreDistributionData}
-          options={barChartOptions}
-        />
-      </div>
-      {/* <div className={styles.processHeadingChart}>
-        <h3>Normality Analysis Data:</h3>
-      </div>
-      <Bar data={normalityAnalysisData} options={barChartOptions}/>
-      <br />
+//       <div style={{ display: 'flex', flexDirection: 'row', justifyContent: 'space-between' }}>
 
-      <div className={styles.processHeadingChart}>
-        <h3>Bar Chart for Central Tendencies:</h3>
-      </div>
-      <Bar data={barChartData} options={barChartOptions}/>
-      <br />
+//         <Pie
+//           data={pieChartData}
+//           options={barChartOptions}
+//         />
+//         <Pie
+//           data={npsScoreDistributionData}
+//           options={barChartOptions}
+//         />
+//       </div>
+//       {/* <div className={styles.processHeadingChart}>
+//         <h3>Normality Analysis Data:</h3>
+//       </div>
+//       <Bar data={normalityAnalysisData} options={barChartOptions}/>
+//       <br />
 
-      <div className={styles.processHeadingChart}>
-        <h3>Pie Chart for Score List:</h3>
-      </div>
-      <Pie
-        data={pieChartData}
-        options={barChartOptions}
-      />
-      <br />
+//       <div className={styles.processHeadingChart}>
+//         <h3>Bar Chart for Central Tendencies:</h3>
+//       </div>
+//       <Bar data={barChartData} options={barChartOptions}/>
+//       <br />
 
-      <div className={styles.processHeadingChart}>
-        <h3>Pie Chart for NPS Score Distribution:</h3>
-      </div>
-      <Pie
-        data={npsScoreDistributionData}
-        options={barChartOptions}
-      /> */}
-    </div>
-  );
-};
+//       <div className={styles.processHeadingChart}>
+//         <h3>Pie Chart for Score List:</h3>
+//       </div>
+//       <Pie
+//         data={pieChartData}
+//         options={barChartOptions}
+//       />
+//       <br />
+
+//       <div className={styles.processHeadingChart}>
+//         <h3>Pie Chart for NPS Score Distribution:</h3>
+//       </div>
+//       <Pie
+//         data={npsScoreDistributionData}
+//         options={barChartOptions}
+//       /> */}
+//     </div>
+//   );
+// };
 
 export default ProcessDetail
