+++ src/pages/Processes/AllProccessPage/ProcessesPage.js
@@ -20,6 +20,7 @@ import {
 import { useNavigate } from 'react-router-dom';
 import { productName } from '../../../utils/helpers';
 import { useAppContext } from '../../../contexts/AppContext';
+import EvaluationReportComponent from '../../../components/manageFiles/ProcessDetail/StepDetail';
 
 const ProcessesPage = ({
   home,
@@ -31,6 +32,7 @@ const ProcessesPage = ({
   showOnlyTests,
   showOnlyCompleted,
   showOnlyActive,
+  showEvaluationReport
 }) => {
   const {
     processesLoading,
@@ -61,6 +63,7 @@ const ProcessesPage = ({
   useEffect(() => {
     if (showOnlySaved) navigate('#saved-processes');
     if (showSingleProcess) navigate('#processdetail');
+    if (showEvaluationReport) navigate('#evaluation-report');
     if (showOnlyPaused) navigate('#paused-processes');
     if (showOnlyCancelled) navigate('#cancelled-processes');
     if (showOnlyTrashed) navigate('#thrashed-processes');
@@ -97,13 +100,13 @@ const ProcessesPage = ({
     const [userCompanyId, userPortfolioDataType] = [
       userDetail?.portfolio_info?.length > 1
         ? userDetail?.portfolio_info.find(
-            (portfolio) => portfolio.product === productName
-          )?.org_id
+          (portfolio) => portfolio.product === productName
+        )?.org_id
         : userDetail?.portfolio_info[0]?.org_id,
       userDetail?.portfolio_info?.length > 1
         ? userDetail?.portfolio_info.find(
-            (portfolio) => portfolio.product === productName
-          )?.data_type
+          (portfolio) => portfolio.product === productName
+        )?.data_type
         : userDetail?.portfolio_info[0]?.data_type,
     ];
 
@@ -138,8 +141,8 @@ const ProcessesPage = ({
     const userPortfolioDataType =
       userDetail?.portfolio_info?.length > 1
         ? userDetail?.portfolio_info.find(
-            (portfolio) => portfolio.product === productName
-          )?.data_type
+          (portfolio) => portfolio.product === productName
+        )?.data_type
         : userDetail?.portfolio_info[0].data_type;
 
     setCurrentUserPortfolioDataType(userPortfolioDataType);
@@ -204,6 +207,14 @@ const ProcessesPage = ({
           {showSingleProcess ? (
             <div id='processdetail'>
               <ProcessDetail />
+            </div>
+          ) : (
+            <></>
+          )}
+
+          {showEvaluationReport ? (
+            <div id='evaluation'>
+              <EvaluationReportComponent />
             </div>
           ) : (
             <></>
